Stabilize HNSW bulk ingestion for 0.2.0-rc2
- Implement single-clone GraphData strategy for bulk operations to eliminate COW overhead.
- Add EnsureChunks for atomic multi-chunk pre-allocation.
- Fix race conditions and deadlocks in concurrent AddBatch via robust nodeCount synchronization.
- Resolve G115 gosec integer overflow warnings in GPU/TPU stubs.
- Ensure ARM64 NEON SIMD kernels build and pass all accuracy tests.

+++ internal/gpu/cuda/cuda_backend_linux.go
@@ -24,8 +24,8 @@ type CUDADevice struct {
 
 // InitializeCUDA initializes the CUDA runtime for the specified device
 func InitializeCUDA(deviceID int) error {
-	if deviceID < 0 || deviceID > math.MaxInt32 { // #nosec G115
-		return fmt.Errorf("invalid deviceID: %d", deviceID) // #nosec G115
+	if deviceID < 0 || int64(deviceID) > int64(math.MaxInt32) {
+		return fmt.Errorf("invalid deviceID: %d", deviceID)
 	}
 	devIDi32 := int32(deviceID) // #nosec G115
 	ret := C.lb_cuda_init_device(C.int(devIDi32))
